requirements.txt
dbt-duckdb==1.10.0
dbt-core==1.11.2